+++ src/azure-cli-core/HISTORY.rst
@@ -3,10 +3,14 @@
 Release History
 ===============
 
+2.0.32
+++++++
+* Minor fixes
 
 2.0.31
 ++++++
 * Allow other sources to add additional tab completion choices via event hook
+* `sdist` is now compatible with wheel 0.31.0
 
 2.0.30
 ++++++