+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -3,6 +3,15 @@
 Release History
 ===============
 
+2.0.24
+++++++
+* Improve repository commands error messages.
+
+2.0.23
+++++++
+* Improve error handling of wincred fallback.
+* `sdist` is now compatible with wheel 0.31.0
+
 2.0.22
 ++++++
 * Improve repository delete command with --image parameter to support docker image format.