+++ src/command_modules/azure-cli-batch/HISTORY.rst
@@ -3,9 +3,14 @@
 Release History
 ===============
 
+3.2.1
+++++++
+* Minor fixes
+
 3.2.0
 ++++++
 * Updated to Batch SDK 4.1.1.
+* `sdist` is now compatible with wheel 0.31.0
 
 3.1.11
 ++++++