+++ src/command_modules/azure-cli-find/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+0.2.9
+++++++
+
+* `sdist` is now compatible with wheel 0.31.0
+
 0.2.8
 ++++++
 * Update for CLI core changes.