+++ src/command_modules/azure-cli-iot/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+0.1.19
+++++++
+
+* `sdist` is now compatible with wheel 0.31.0
+
 0.1.18
 ++++++
 * Support Autorest 3.0 based SDKs